Moved MapArea-related things to its own file, minor fixes
