+++ DS2S META/TabControls/RandomizerSettings.xaml
@@ -5,13 +5,14 @@
                     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
                     xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
                     xmlns:local="clr-namespace:DS2S_META"
+                    xmlns:rando="clr-namespace:DS2S_META.Randomizer"
                     mc:Ignorable="d" 
                     d:DesignHeight="600" d:DesignWidth="425" Background="LightGray">
 
     <StackPanel>
         <Expander>
             <Expander.Header>
-                <TextBlock FontSize="16" Margin="4" FontWeight="Bold">Specific Item Restrictions</TextBlock>
+                <TextBlock FontSize="16" Margin="4" FontWeight="Bold">Item Placement Restrictions</TextBlock>
             </Expander.Header>
             <ListView x:Name="listViewItemRestrictions" ItemsSource="{Binding RelativeSource={RelativeSource AncestorType={x:Type local:RandomizerSettings}}, Path=ItemRestrictions, UpdateSourceTrigger=PropertyChanged, Mode=TwoWay}" HorizontalContentAlignment="Stretch">
                 <ListView.ItemTemplate>
@@ -67,7 +68,7 @@
                     <ColumnDefinition Width="4*"/>
                 </Grid.ColumnDefinitions>
                 <TextBlock HorizontalAlignment="Right" VerticalAlignment="Center" TextAlignment="Right" FontSize="14" Text="From:" Margin="0,0,0,4"/>
-                <ComboBox x:Name="connectedAreaDistanceComboBox" Grid.Column="1" HorizontalAlignment="Stretch" Margin="6" DisplayMemberPath="Value" FontSize="14" ItemsSource="{Binding Source={x:Static local:RandomizerSettings.MapAreaToString}}" SelectionChanged="ConnectedAreaDistanceComboBox_SelectionChanged"/>
+                <ComboBox x:Name="connectedAreaDistanceComboBox" Grid.Column="1" HorizontalAlignment="Stretch" Margin="6" DisplayMemberPath="Value" FontSize="14" ItemsSource="{Binding Source={x:Static rando:MapAreas.toString}}" SelectionChanged="ConnectedAreaDistanceComboBox_SelectionChanged"/>
                 <ListView x:Name="listViewConnectedAreasDistances" Grid.Row="1" Grid.ColumnSpan="2" HorizontalContentAlignment="Stretch" ItemsSource="{Binding RelativeSource={RelativeSource AncestorType={x:Type local:RandomizerSettings}}, Path=ConnectedAreaDistanceListItems, UpdateSourceTrigger=PropertyChanged, Mode=TwoWay}" Visibility="Collapsed">
                     <ListView.ItemTemplate>
                         <DataTemplate>
